issue#3: Changed interfaces of functions SPP_read and SPP_write to accept uint32_t params instead of uint16_t in order to reflect the HW register width. Added doxygen docu.

+++ lpc1343_spp/inc/spp.h
@@ -9,21 +9,25 @@
 
 #include "board.h"
 
+/**
+ *  @enum   SPP_WORD_SIZE
+ *  @brief  The enums define the word size for the transfer.
+ */
 typedef enum
 {
-	SPP_WORD_SIZE_04_BIT = 0x3
-,	SPP_WORD_SIZE_05_BIT = 0x4
-,	SPP_WORD_SIZE_06_BIT = 0x5
-,	SPP_WORD_SIZE_07_BIT = 0x6
-,	SPP_WORD_SIZE_08_BIT = 0x7
-,	SPP_WORD_SIZE_09_BIT = 0x8
-,	SPP_WORD_SIZE_10_BIT = 0x9
-,	SPP_WORD_SIZE_11_BIT = 0xA
-,	SPP_WORD_SIZE_12_BIT = 0xB
-,	SPP_WORD_SIZE_13_BIT = 0xC
-,	SPP_WORD_SIZE_14_BIT = 0xD
-,	SPP_WORD_SIZE_15_BIT = 0xE
-,	SPP_WORD_SIZE_16_BIT = 0xF
+    SPP_WORD_SIZE_04_BIT = 0x3
+,   SPP_WORD_SIZE_05_BIT = 0x4
+,   SPP_WORD_SIZE_06_BIT = 0x5
+,   SPP_WORD_SIZE_07_BIT = 0x6
+,   SPP_WORD_SIZE_08_BIT = 0x7
+,   SPP_WORD_SIZE_09_BIT = 0x8
+,   SPP_WORD_SIZE_10_BIT = 0x9
+,   SPP_WORD_SIZE_11_BIT = 0xA
+,   SPP_WORD_SIZE_12_BIT = 0xB
+,   SPP_WORD_SIZE_13_BIT = 0xC
+,   SPP_WORD_SIZE_14_BIT = 0xD
+,   SPP_WORD_SIZE_15_BIT = 0xE
+,   SPP_WORD_SIZE_16_BIT = 0xF
 } SPP_WORD_SIZE;
 
 /** The SPP shall communicate using the SPI format. */
@@ -64,6 +68,10 @@ typedef enum
 /** Bitmask for DIV in SSP0CLKDIV. */
 #define SPP_DIV_MASK                ((uint32_t)0x000000FF)
 
+/**
+ *  @struct SPP_STATUS
+ *  @brief  Structure SPP_STATUS reflects the SSP0SSR register.
+ */
 typedef struct
 {
     Bool TFE;
@@ -74,14 +82,14 @@ typedef struct
 } SPP_STATUS;
 
 /**
- *  @brief  Function SPI_init configures the LPC's SPI unit.
- *  @param  frf Defines the frame format (SPI, TI, MICRO)
- *  @param  cpol Defines the clock out polarity.
- *  @param  cpha Defines the clock out phase.
- *  @param  scr Defines the serial clock rate.
- *  @param  prescale_val Defines the prescaler value of register CPSDVSR.
- *  @param  clk_div Defines the clock divider value of register DIV.
- *  @return void
+ *  @brief      Function SPP_init configures the LPC's SPP unit.
+ *  @param[in]  frf Defines the frame format (SPI, TI, MICRO)
+ *  @param[in]  cpol Defines the clock out polarity.
+ *  @param[in]  cpha Defines the clock out phase.
+ *  @param[in]  scr Defines the serial clock rate.
+ *  @param[in]  prescale_val Defines the prescaler value of register CPSDVSR.
+ *  @param[in]  clk_div Defines the clock divider value of register DIV.
+ *  @return     void
  */
 void SPP_init ( SPP_WORD_SIZE dss
               , uint8_t frf
@@ -92,10 +100,29 @@ void SPP_init ( SPP_WORD_SIZE dss
               , uint8_t clk_div
               );
 
-void SPP_write (uint16_t data);
+/**
+ *  @brief      Function SPP_write sends a 16-bit word to the SPP.
+ *  @param[in]  data the data to be written. The high word will be masked.
+ *  @return     void
+ */
+void SPP_write (uint32_t data);
 
-uint16_t SPP_read ();
+/**
+ *  @brief      Function SPP_read reads a 16-bit word from the SPP.
+ *  @return     uint32_t
+ *  @retval     The 16-bit word from the SPP.
+ */
+uint32_t SPP_read ();
 
+/**
+ *  @brief      Function SPP_get_status gets the status of the SPP from register
+ *              <b>SSP0SR</b>.
+ *  @param[in]  status a pointer to a structure <b>SPP_STATUS</b>, where the
+ *              regiter's values will be stored to.
+ *  @return     Bool
+ *  @retval     FALSE Retrieving status failed.
+ *  @retval     TRUE Retrieving status succeeded.
+ */
 Bool SPP_get_status (SPP_STATUS *status);
 
 #endif /* INC_SPP_H_ */
